fix(frontend): move layout folder location
